# openapi.yaml
openapi: 3.1.0
info:
  title: Document Management API
  description: API for managing document uploads, downloads, and extractions
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: /api
    description: API base URL
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              description: Error type identifier
              enum:
                - VALIDATION_ERROR
                - NOT_FOUND
                - UNAUTHORIZED
                - FORBIDDEN
                - CONFLICT
                - INTERNAL_SERVER_ERROR
                - BAD_REQUEST
                - SERVICE_UNAVAILABLE
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
      required:
        - error
  parameters:
    DocumentIdParam:
      name: id
      in: path
      required: true
      description: Document ID
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
paths:
...
  /documents/{id}/download-url:
    parameters:
      - $ref: "#/components/parameters/DocumentIdParam"

    get:
      summary: Get a download URL for a document
      description: |
        Generates a pre-signed URL for downloading or viewing a document.

        The URL is valid for a limited time, specified by the expiresIn parameter.
      parameters:
        - name: expiresIn
          in: query
          description: URL expiration time in seconds (60-604800)
          schema:
            type: integer
            default: 3600
            minimum: 60
            maximum: 604800
        - name: detailed
          in: query
          description: Whether to return a detailed response
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Download URL generated successfully
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    description: Simple response (legacy format)
                    properties:
                      url:
                        type: string
                        format: uri
                        description: Pre-signed URL for downloading the document
                  - type: object
                    description: Detailed response
                    properties:
                      url:
                        type: string
                        format: uri
                        description: Pre-signed URL for downloading the document
                      expires:
                        type: string
                        format: date-time
                        description: Expiration time of the URL
                      documentId:
                        type: string
                        format: uuid
                        description: Document ID
                      documentName:
                        type: string
                        description: Document name
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...